chowfast_backend/requirements.txt
django==5.2.8
isort==7.0.0
python-dotenv==1.2.1
python-decouple==3.8
psycopg2-binary==2.9.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1 
drf-yasg==1.21.11
argon2-cffi==25.1.0
django-cors-headers==4.9.0
gunicorn==23.0.0
brotli==1.2.0 
whitenoise==6.11.0
uvicorn==0.38.0
django-jazzmin==3.0.1
django-anymail==13.1
Pillow==12.0.0 
cloudinary==1.44.1 
django-cloudinary-storage==0.3.0 